big refractor
The actions that a user can choose are now in a folder, with each action a different file.
