+++ functions/functionalities/exit_app.py
@@ -1,0 +1,3 @@
+def exit_app():
+    print("Exiting...")
+    exit()
